x86/sev: Move sev_guest.h into common SEV header
jira LE-2290
Rebuild_History Non-Buildable kernel-5.14.0-503.22.1.el9_5
commit-author Michael Roth <michael.roth@amd.com>
commit f55f3c3ac69f22ce092506244a31c08a1ca497ba

sev_guest.h currently contains various definitions relating to the
format of SNP_GUEST_REQUEST commands to SNP firmware. Currently only the
sev-guest driver makes use of them, but when the KVM side of this is
implemented there's a need to parse the SNP_GUEST_REQUEST header to
determine whether additional information needs to be provided to the
guest. Prepare for this by moving those definitions to a common header
that's shared by host/guest code so that KVM can also make use of them.

Message-ID: <20240701223148.3798365-3-michael.roth@amd.com>
(cherry picked from commit f55f3c3ac69f22ce092506244a31c08a1ca497ba)
